Remove the offering (Issue Note) feature and fold its shared code into Community Credit
Community Credit is the new UI/UX built on the same FixedReturn.sol contract, making
the old offering views redundant. Deletes the FixedReturnView components/views/routes/
nav entry, and relocates every offering-named util, type, and schema that Community
Credit still depends on into properly-named Community Credit / FixedReturn files
(FixedReturn.sol is Community Credit's own contract, so there's no other consumer left
to justify a separate generic layer). Also adds test coverage for StepIndicator.vue,
which had none even before this move.
