+++ app/src/components/ui/StepIndicator.vue
