+++ app/src/types/communityCredit.ts
@@ -2,11 +2,100 @@
 // Community Credit lets a company raise working capital from its members:
 // each round is repaid in full with fixed interest at maturity, signed on-chain
 // from a dedicated Credit Account. This module is fake-data only (no API yet).
+// FixedReturn.sol is Community Credit's own contract — every on-chain shape below
+// mirrors it directly.
 
-import type { WhitelistEntry } from './offering'
+import type { Address } from 'viem'
 import type { UBadgeColor } from './ui'
 
 export type CreditRole = 'owner' | 'lender'
+
+/** FixedReturn.sol's TermUnit enum (Days, Months, Years), as a string union. */
+export type FixedReturnTermUnit = 'days' | 'months' | 'years'
+
+/** FixedReturn.sol's FundingAccess enum (General, Whitelist), as a string union. */
+export type FixedReturnAccessMode = 'general' | 'whitelist'
+
+/** A single whitelisted lender entry for a FixedReturn offer's CreateOfferParams. */
+export interface FixedReturnWhitelistEntry {
+  username: string
+  address: string
+  amount: number | null
+}
+
+/**
+ * Matches FixedReturn.sol's `CreateOfferParams` struct field-for-field — the on-chain
+ * args for `createLendingOffer`.
+ */
+export interface FixedReturnOfferParams {
+  token: Address
+  fundingTarget: bigint
+  interestRateBps: bigint
+  termDuration: number
+  termUnit: 0 | 1 | 2
+  startDate: bigint
+  subscriptionDeadline: bigint
+  fundingAccess: 0 | 1
+  isCapEnabled: boolean
+  lenderCap: bigint
+  whitelistAddrs: Address[]
+  allocations: bigint[]
+}
+
+/** Raw shape decoded from FixedReturn.sol's getLendingOffer (named-tuple struct). */
+export interface LendingOfferStruct {
+  token: Address
+  fundingTarget: bigint
+  interestRateBps: bigint
+  termDuration: number
+  termUnit: 0 | 1 | 2
+  startDate: bigint
+  subscriptionDeadline: bigint
+  fundingAccess: 0 | 1
+  isCapEnabled: boolean
+  lenderCap: bigint
+  totalFunded: bigint
+  totalRepaidByIssuer: bigint
+  state: 0 | 1 | 2 | 3
+}
+
+/** A single FixedReturn offer as returned by useFixedReturnAllOffers. */
+export interface FixedReturnRawOffer {
+  offerId: number
+  offer: LendingOfferStruct
+  decimals: number
+}
+
+/** A single lender's position on one offer, as returned by useFixedReturnOfferLenders. */
+export interface FixedReturnOfferLender {
+  address: Address
+  principal: number
+  expected: number
+}
+
+/**
+ * A connected lender's position (Whitelist allocation + cumulative deposits) on one
+ * offer, as returned by useFixedReturnMyLenderPositions.
+ */
+export interface FixedReturnLenderPosition {
+  allocation: bigint
+  deposited: bigint
+}
+
+/**
+ * Off-chain title/purpose metadata for a FixedReturn offer, as returned by
+ * GET /api/fixed-return-offering — see backend/prisma/schema.prisma's
+ * FixedReturnOffering model.
+ */
+export interface FixedReturnOfferingResponse {
+  id: number
+  teamId: number
+  offerId: number
+  title: string
+  purpose: string | null
+  createdAt: string
+  updatedAt: string
+}
 
 // Mirrors FixedReturn.sol's OfferState (Open, Funded, Refundable, Repaying) resolved
 // against the offer's deadline and repayment progress. 'active' = Repaying but not yet
@@ -27,11 +116,11 @@ export interface StatusMeta {
 
 export type CreditAccess = 'everyone' | 'restricted'
 
-/** Community Credit's own whitelist entry — same shape as Issue Note's WhitelistEntry
+/** Community Credit's own whitelist entry — same shape as a FixedReturnWhitelistEntry
  *  plus `custom`. A lender defaults to tracking the round-level cap live (amount stays
  *  in sync whenever the cap changes); `custom: true` means their amount was set
  *  independently and stops following further cap changes. */
-export interface CreditWhitelistEntry extends WhitelistEntry {
+export interface CreditWhitelistEntry extends FixedReturnWhitelistEntry {
   custom?: boolean
 }
 
@@ -55,13 +144,12 @@ export interface CreditLender {
   /** Two comma-separated hex colors used to render the avatar gradient. */
   gradient: string
   amount: number
-  /** Total owed at maturity — principal + interest (mirrors Issue Note's "Expected
-   *  return" column, from FixedReturn's totalEntitlementOf). */
+  /** Total owed at maturity — principal + interest, from FixedReturn's
+   *  totalEntitlementOf. */
   expected: number
   /** Paid to this lender so far — proportional estimate (this lender's share of
-   *  raised × the round's totalRepaid), mirroring Issue Note's buildFixedReturnLenderRows
-   *  rather than an extra per-lender contract read (totalPaidToLender exists on-chain but
-   *  Issue Note doesn't call it either, for the same reason). */
+   *  raised × the round's totalRepaid) rather than an extra per-lender contract read
+   *  (totalPaidToLender exists on-chain but isn't called, for the same reason). */
   paid: number
   /** Short, human display date, e.g. `Jun 3`. */
   date: string
@@ -129,4 +217,53 @@ export interface CreditCallForm {
   whitelist: CreditWhitelistEntry[]
   capOn: boolean
   cap: string
+}
+
+/**
+ * The on-chain-ready shape a credit round's Basics/Terms/Access wizard fields resolve
+ * to before being handed to toFixedReturnOfferParams — distinct from CreditCallForm,
+ * which is the raw wizard state (string-typed fields, preset/custom period picker).
+ */
+export interface CreditOfferForm {
+  title: string
+  purpose: string
+  principal: number
+  rate: number
+  termValue: number
+  termUnit: FixedReturnTermUnit
+  deadline: string
+  access: FixedReturnAccessMode
+  capOn: boolean
+  cap: number
+  token: string | undefined
+}
+
+/**
+ * A FixedReturn offer from the lender's point of view — eligibility and the deposit
+ * cap depend on the connected wallet. There's no on-chain "minimum deposit" or "fixed amount" concept —
+ * `cap` is either the General-mode per-lender lenderCap, the Whitelist-mode personal
+ * allocation, or null (no cap, General mode with isCapEnabled false). `remaining` is
+ * the amount the lender can currently submit: the smaller of their remaining personal
+ * limit (when one exists) and the offer-wide room left before the funding target.
+ */
+export interface CreditLenderOffering {
+  id: string
+  title: string
+  rate: number
+  term: number
+  termUnit: FixedReturnTermUnit
+  access: FixedReturnAccessMode
+  allowed: boolean
+  cap: number | null
+  remaining: number
+  myDeposited: number
+  raised: number
+  target: number
+  token: Address
+  accessLabel: string
+  accessBg: string
+  accessColor: string
+  accessDot: string
+  limitsLabel: string
+  pct: number
 }
