+++ app/src/types/index.ts
@@ -1,6 +1,4 @@
 export * from './ui'
-export * from './offering'
-export * from './offering.schemas'
 export * from './communityCredit'
 export * from './communityCredit.schemas'
 export * from './toast-type'
