+++ app/src/utils/index.ts
@@ -1,5 +1,4 @@
-export * from './offeringUtil'
-export * from './offeringViewUtil'
+export * from './communityCreditOfferUtil'
 export * from './communityCreditUtil'
 export * from './errorUtil'
 export * from './classifyError'
