+++ app/src/views/team/[id]/CommunityCredit/NewView.vue
@@ -162,11 +162,11 @@ import {
   applyZodFieldErrors,
   classifyError,
   creditChipClass,
-  getSupportedOfferingTokenOptions,
+  getSupportedCreditTokenOptions,
   toCreditCallOfferParams
 } from '@/utils'
-import { creditCallBasicsSchema, type CreditCallForm, type OfferingForm } from '@/types'
-import StepIndicator from '@/components/sections/FixedReturnView/StepIndicator.vue'
+import { creditCallBasicsSchema, type CreditCallForm, type CreditOfferForm } from '@/types'
+import StepIndicator from '@/components/ui/StepIndicator.vue'
 import CreditCallAccessStep from '@/components/sections/CommunityCreditView/CreditCallAccessStep.vue'
 import CreditCallTermsStep from '@/components/sections/CommunityCreditView/CreditCallTermsStep.vue'
 import CreditCallSummaryCard from '@/components/sections/CommunityCreditView/CreditCallSummaryCard.vue'
@@ -190,7 +190,7 @@ const accessStepRef = ref<StepHandle>(null)
 // Only tokens this team's FixedReturn contract actually accepts (ERC20-only).
 const { data: supportedTokens } = useFixedReturnGetSupportedTokens()
 const tokens = computed(() =>
-  getSupportedOfferingTokenOptions((supportedTokens.value as Address[] | undefined) ?? []).map(
+  getSupportedCreditTokenOptions((supportedTokens.value as Address[] | undefined) ?? []).map(
     (option) => option.value
   )
 )
@@ -269,7 +269,7 @@ async function publish() {
     // the subscription deadline, so startDate == deadline. FixedReturn.sol requires
     // subscriptionDeadline <= startDate (reverts InvalidDeadline otherwise). The term is
     // already in canonical days, so it maps straight to the contract's Days unit.
-    const offeringForm: OfferingForm = {
+    const offeringForm: CreditOfferForm = {
       title: form.name.trim(),
       purpose: form.desc.trim(),
       principal: Number(form.target) || 0,
