--- app/src/views/team/[id]/LenderMarketplaceView.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class="flex flex-col gap-6">
-    <LenderMarketplace />
-  </div>
-</template>
-
-<script setup lang="ts">
-import LenderMarketplace from '@/components/sections/FixedReturnView/LenderMarketplace.vue'
-</script>
